+++ src/Oro/Bundle/FormBundle/Form/Extension/StripTagsExtension.php
@@ -4,27 +4,42 @@ namespace Oro\Bundle\FormBundle\Form\Extension;
 
 use Oro\Bundle\FormBundle\Form\Extension\Traits\FormExtendedTypeTrait;
 use Oro\Bundle\UIBundle\Tools\HtmlTagHelper;
+use Psr\Container\ContainerInterface;
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\FormEvent;
 use Symfony\Component\Form\FormEvents;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Contracts\Service\ServiceSubscriberInterface;
 
-class StripTagsExtension extends AbstractTypeExtension
+/**
+ * Strips all HTML tags in submitted data if "strip_tags" option is set.
+ */
+class StripTagsExtension extends AbstractTypeExtension implements ServiceSubscriberInterface
 {
     use FormExtendedTypeTrait;
 
     const OPTION_NAME = 'strip_tags';
 
-    /** @var HtmlTagHelper */
-    protected $htmlTagHelper;
+    /** @var ContainerInterface */
+    private $container;
+
+    /**
+     * @param ContainerInterface $container
+     */
+    public function __construct(ContainerInterface $container)
+    {
+        $this->container = $container;
+    }
 
     /**
      * {@inheritDoc}
      */
-    public function __construct(HtmlTagHelper $htmlTagHelper)
+    public static function getSubscribedServices()
     {
-        $this->htmlTagHelper = $htmlTagHelper;
+        return [
+            'oro_ui.html_tag_helper' => HtmlTagHelper::class
+        ];
     }
 
     /**
@@ -55,7 +70,9 @@ class StripTagsExtension extends AbstractTypeExtension
         return function (FormEvent $event) {
             $data = $event->getData();
             if (is_string($data)) {
-                $event->setData($this->htmlTagHelper->stripTags($data));
+                /** @var HtmlTagHelper $htmlTagHelper */
+                $htmlTagHelper = $this->container->get('oro_ui.html_tag_helper');
+                $event->setData($htmlTagHelper->stripTags($data));
             }
         };
     }