+++ src/Oro/Bundle/NoteBundle/Provider/NoteActivityListProvider.php
@@ -78,14 +78,6 @@ class NoteActivityListProvider implements
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function getAclClass()
-    {
-        return null;
-    }
-
-    /**
      * @param Note $entity
      *
      * {@inheritdoc}